Prevent purging of remote data on server start unless specifically requested

+++ src/main/java/org/buddycloud/channelserver/sync/ServerSync.java
@@ -4,6 +4,7 @@ import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 
 import org.apache.log4j.Logger;
+import org.buddycloud.channelserver.Configuration;
 import org.buddycloud.channelserver.channel.ChannelManager;
 import org.buddycloud.channelserver.channel.ChannelManagerFactory;
 import org.buddycloud.channelserver.db.exception.NodeStoreException;
@@ -16,26 +17,34 @@ public class ServerSync {
 	private ChannelManager channelManager;
 	private BlockingQueue<Packet> inQueue;
 	private BlockingQueue<Packet> outQueue;
+	private Configuration configuration;
 
 	private Logger logger = Logger.getLogger(ServerSync.class);
-	
+
 	public ServerSync(ChannelManagerFactory channelManagerFactory,
-			BlockingQueue<Packet> outQueue, BlockingQueue<Packet> inQueue) {
+			BlockingQueue<Packet> outQueue, BlockingQueue<Packet> inQueue,
+			Configuration configuration) {
 		this.channelManager = channelManagerFactory.create();
 		this.outQueue = outQueue;
 		this.inQueue = inQueue;
+		this.configuration = configuration;
 	}
 
 	public void start() {
 		try {
-		    deleteFederatedChannelCache();
-		    channelManager.close();
+			deleteFederatedChannelCache();
+			channelManager.close();
 		} catch (NodeStoreException e) {
 			logger.error(e);
 		}
 	}
 
 	private void deleteFederatedChannelCache() throws NodeStoreException {
+		Boolean purge = Boolean.valueOf(configuration.getProperty(
+				Configuration.PURGE_REMOTE_ON_START, "false"));
+		if (false == purge) {
+			return;
+		}
 		channelManager.deleteRemoteData();
 	}
 }
